+++ src/services/notification/index.js
@@ -1,10 +1,10 @@
+// eslint-disable-next-line max-classes-per-file
 const request = require('request-promise-native');
 const { static: staticContent } = require('@feathersjs/express');
 const path = require('path');
 
+const { Configuration } = require('@schul-cloud/commons');
 const hooks = require('./hooks/index');
-
-const { REQUEST_TIMEOUT } = require('../../../config/globals');
 
 /**
  * maps jsonapi properties of a response to fit anything but jsonapi
@@ -38,7 +38,7 @@ class MessageService {
 			},
 			body: Object.assign(data, { serviceUrl: serviceUrls.notification }),
 			json: true,
-			timeout: REQUEST_TIMEOUT,
+			timeout: Configuration.get('REQUEST_TIMEOUT'),
 		};
 
 		return request(options).then((response) => response);
@@ -54,7 +54,7 @@ class MessageService {
 				token: userId,
 			},
 			json: true,
-			timeout: REQUEST_TIMEOUT,
+			timeout: Configuration.get('REQUEST_TIMEOUT'),
 		};
 
 		return request(options).then((message) => message);
@@ -80,7 +80,7 @@ class DeviceService {
 				token: userId,
 			},
 			json: true,
-			timeout: REQUEST_TIMEOUT,
+			timeout: Configuration.get('REQUEST_TIMEOUT'),
 		};
 
 		return request(options).then((devices) => devices);
@@ -99,7 +99,7 @@ class DeviceService {
 			},
 			body: data,
 			json: true,
-			timeout: REQUEST_TIMEOUT,
+			timeout: Configuration.get('REQUEST_TIMEOUT'),
 		};
 
 		return request(options).then((response) => mapResponseProps(response));
@@ -113,7 +113,7 @@ class DeviceService {
 			uri: `${serviceUrls.notification}/devices/${id}?token=${userId}`,
 			method: 'DELETE',
 			json: true,
-			timeout: REQUEST_TIMEOUT,
+			timeout: Configuration.get('REQUEST_TIMEOUT'),
 		};
 
 		return request(options).then((message) => message);
@@ -141,7 +141,7 @@ class CallbackService {
 			},
 			body: data,
 			json: true,
-			timeout: REQUEST_TIMEOUT,
+			timeout: Configuration.get('REQUEST_TIMEOUT'),
 		};
 
 		return request(options).then((response) => response);
@@ -167,7 +167,7 @@ class NotificationService {
 				token: userId,
 			},
 			json: true,
-			timeout: REQUEST_TIMEOUT,
+			timeout: Configuration.get('REQUEST_TIMEOUT'),
 		};
 
 		return request(options).then((message) => message);
@@ -184,7 +184,7 @@ class NotificationService {
 				token: userId,
 			},
 			json: true,
-			timeout: REQUEST_TIMEOUT,
+			timeout: Configuration.get('REQUEST_TIMEOUT'),
 		};
 
 		return request(options).then((message) => message);